Rearranged components into separate files.
